read/pe: split into modules
